requirements.txt
dataclasses==0.6
pathlib==1.0.1
pywin32==304
winshell==0.6
vdf==3.4
pyyaml==6.0
pytest==7.1.2
pytest-cov==3.0.0
pyinstaller==5.2
